#13- Added ReplaceCharacters method and refactored unit tests.
